move verify_approvals_and_threshold_orphan to EMA
This is slighly less trivial than pure move -- the API is adjusted to
remove direct dependency on doomslug which lives in a different crate
